feat: increase token coverage to output full schema for scss & js, refactor build files
