+++ src/properties/sizes.json
@@ -18,6 +18,18 @@
       "4": { "value": 2 },
       "5": { "value": 4 },
       "6": { "value": 8 }
+    },
+    "size": {
+      "0": { "value": 1 },
+      "1": { "value": 2 },
+      "2": { "value": 3 },
+      "3": { "value": 4 }
+    },
+    "radii": {
+      "0": { "value": 0.25 },
+      "1": { "value": 0.5 },
+      "2": { "value": 0.75 },
+      "3": { "value": 1 }
     }
   }
 }
